+++ app/Werewolf/Options.hs
@@ -19,6 +19,7 @@ module Werewolf.Options (
     werewolfPrefs, werewolfInfo, werewolf,
 ) where
 
+import Data.List.Extra
 import           Data.Text    (Text)
 import qualified Data.Text    as T
 import           Data.Version (showVersion)
@@ -107,7 +108,7 @@ see = See . See.Options . T.pack <$> strArgument (metavar "PLAYER")
 
 start :: Parser Command
 start = fmap Start $ Start.Options
-    <$> fmap (T.splitOn "," . T.pack) (strOption $ mconcat [
+    <$> fmap (map T.pack . wordsBy (',' ==)) (strOption $ mconcat [
         long "extra-roles", metavar "ROLE,...",
         value [],
         help "Specify the extra roles to include"